+++ spec/LogStore.spec.coffee
@@ -1,7 +1,7 @@
 LogStore = require '../lib/LogStore'
 LogPublisher = require '../lib/LogPublisher'
-{createServer} = require './util/redis-remote'
-{isWindows, notyet,mocklog,runUntil} = require './util/SpecHelpers'
+MockRedis = require './util/MockRedis'
+{isWindows,notyet,mocklog,runUntil} = require './util/SpecHelpers'
 redis = require 'redis'
 
 if isWindows() and (not process.env.TEST_REDIS_HOST or not process.env.TEST_REDIS_PORT)
@@ -13,37 +13,22 @@ DEFAULT_COUNT = 25
 describe "LogStore", ->
   ls = null
   db = null
-  redishost = null
-  redisport = null
-  dbmgr = null
+  mockRedis = null
   curdbid = 0
 
   beforeEach ->
-    if not db
-      # Windows/Redis on remote machine
-      if process.env.TEST_REDIS_HOST and process.env.TEST_REDIS_PORT
-        redishost = process.env.TEST_REDIS_HOST
-        redisport = process.env.TEST_REDIS_PORT
-        db = redis.createClient redisport, redishost
-        ls = new LogStore host:redishost, port: redisport
-
-      # Linux: use redis-remote to start and down
-      else
-        redishost = '127.0.0.1'
-        runUntil (done)->
-          createServer ((err,mock)->
-            dbmgr = mock
-            redisport = dbmgr.port
-            db = redis.createClient redisport, redishost
-            ls = new LogStore host:redishost, port: redisport
-            done()
-          ), 100
-    else
-      db.select ++curdbid
-      ls = new LogStore host:redishost, port:redisport, dbid: curdbid
+    runUntil (done)->
+      if not db or not mockRedis
+        new MockRedis (mr)->
+          mockRedis = mr
+          db = redis.createClient mockRedis.port, mockRedis.port
+          done()
+      else done()
+    runs ->
+      ls = new LogStore host:mockRedis.host, port:mockRedis.port, dbid: ++curdbid
+      db.select curdbid
 
   afterEach -> ls?.end()
-
 
   describe ".end()", ->
     it "closes connection, get() sends error to callbacks", ->
@@ -53,9 +38,9 @@ describe "LogStore", ->
 
 
   describe ".on('log', callback)", ->
-    it 'when log, calls callback', ->
-      lp = new LogPublisher context:'test', host:redishost, port:redisport, dbid: curdbid
-      logstores = for i in [0...3] then new LogStore host:redishost, port:redisport, dbid: curdbid
+    it 'calls callback, when Log Entry is published', ->
+      lp = new LogPublisher context:'test', host:mockRedis.host, port:mockRedis.port, dbid: curdbid
+      logstores = for i in [0...3] then new LogStore host:mockRedis.host, port:mockRedis.port, dbid: curdbid
       @after -> for logstore in [lp, logstores...] then logstore.end()
 
       mlogs = [mocklog()]
@@ -73,7 +58,7 @@ describe "LogStore", ->
     mlogs = null
 
     it "returns last #{DEFAULT_COUNT} log entries", ->
-      lp = new LogPublisher context:'test', host:redishost, port:redisport, dbid: curdbid
+      lp = new LogPublisher context:'test', host:mockRedis.host, port:mockRedis.port, dbid: curdbid
       @after -> lp.end()
       mlogs = (mocklog(date: Date.now()+1000*i) for i in [0..DEFAULT_COUNT+5])
 
@@ -84,7 +69,7 @@ describe "LogStore", ->
       runs -> expect(received).toEqual mlogs.reverse().slice(0,DEFAULT_COUNT)
 
     it "returns last X log entries, if X < #{DEFAULT_COUNT}", ->
-      lp = new LogPublisher context:'test', host:redishost, port:redisport, dbid: curdbid
+      lp = new LogPublisher context:'test', host:mockRedis.host, port:mockRedis.port, dbid: curdbid
       @after -> lp.end()
       mlogs = (mocklog(date: Date.now()+1000*i) for i in [0...3])
 
@@ -98,8 +83,8 @@ describe "LogStore", ->
   describe ".get({filterBy: {context}})", ->
 
     it "returns last #{DEFAULT_COUNT} log entries, from specified context", ->
-      lpOne = new LogPublisher context:'one', host:redishost, port:redisport, dbid: curdbid
-      lpTwo = new LogPublisher context:'two', host:redishost, port:redisport, dbid: curdbid
+      lpOne = new LogPublisher context:'one', host:mockRedis.host, port:mockRedis.port, dbid: curdbid
+      lpTwo = new LogPublisher context:'two', host:mockRedis.host, port:mockRedis.port, dbid: curdbid
       @after -> lp.end() for lp in [lpOne, lpTwo]
       oneLogs = (mocklog {category: 'oneCat', date: Date.now()+1000*i} for i in [0...1])
       twoLogs = (mocklog {category: 'twoCat', date: Date.now()+1000*i} for i in [0...1])
@@ -121,7 +106,7 @@ describe "LogStore", ->
     mlogs = undefined
 
     beforeEach ->
-      lp = new LogPublisher context:'test', host:redishost, port:redisport, dbid: curdbid
+      lp = new LogPublisher context:'test', host:mockRedis.host, port:mockRedis.port, dbid: curdbid
       @after -> lp.end()
       mlogs = (mocklog {date: Date.now()+1000*i} for i in [0...5])
       mlogs[0].clientInfo = ip: '1'
@@ -152,7 +137,7 @@ describe "LogStore", ->
     mlogs = undefined
 
     beforeEach ->
-      lp = new LogPublisher context:'test', host:redishost, port:redisport, dbid: curdbid
+      lp = new LogPublisher context:'test', host:mockRedis.host, port:mockRedis.port, dbid: curdbid
       @after -> lp.end()
       mlogs = (mocklog(date: Date.now()+1000*i) for i in [0...5])
       runUntil (done)-> lp.log mlogs, done
@@ -194,9 +179,10 @@ describe "LogStore", ->
       expect(ls._toLog hash).toEqual l
 
 
-  it 'TEST TEARDOWN', ->
-    db.flushall()
-    db.end()
-    dbmgr?.stop()
+  it "MOCK REDIS TEARDOWN", ->
+    if db
+      db.flushall()
+      db.end()
+    mockRedis?.shutdown()
     ls?.end()
     runUntil (done)-> setTimeout done, 100
